Adds route, stop, and arrival models.
Each model has a class `cache_length` attribute. If a record is older
than this amount of time, it should be refreshed from Lextran's
API. Cache lengths are configured in an initializer.

Bus routes have their own IDs assigned by Lextran, so their primary
keys are overridden to be the `route_num` field. Same for stops and
the `stop_num` field.