Adds API models.
Adds Lextran model namespace with Route, Stop, and Arrival models.
Each of these can query the Lextran API for new data.

Arrivals currently don't work. The Lextran bus stop endpoint requires
a referrer and a session in which the user has previously picked
a route. This will be awkward to do on the server.